Add MovementData to PlayerData file
If the server were to go AWOL for any reason during teleportation (say a memory limit being exceeded and the server crashing), the original movement values of a player (which have been set to zero to prevent movement) would be lost.

This would mean on server startup the player would not be able to move at all, and could only hop around the map!

Instead, let’s store this MovementData into the PlayerData file and so should the server go down, then we can restore from this file and everything will be OK.