Rewritten Movement Management Code
The previous versions of this plugin relied on setting the player’s walking and flying speed to 0 in order to keep the player in the same place. While this worked, it caused a major bug in that there are occasions where Spigot doesn’t properly restore a user’s walking and flying speed, meaning the user would be stranded with only the ability to jump.

In the newer versions, we now just set a metadata flag and rely on keeping a player in the same place by using the onPlayerMove() and onPlayerPickup() functions which cancel these events appropriately if a player is wayporting. In this way, we don’t touch setting any flags on a player’s movement.